pyproject.toml
[tool.poetry]
name = "pythonyoutubetranscript"
version = "0.1.0"
description = ""
authors = ["foerster_work <foerster.work@gmail.com>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "<3.13,>=3.9"
youtube-transcript-api = "^1.1.1"
google-genai = "^1.25.0"
python-dotenv = "^1.1.1"
flask = "^3.1.1"

[settings]
[settings.virtualenvs]
in-project = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
